services/orchestrator/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
aiohttp==3.9.1
pydantic>=2.11.5
python-dotenv>=1.0.1
httpx>=0.27.0